warenkorb erstmal fertig; braucht überarbeitung (jeder artikel kann nur exakt 1 mal rein)

+++ code/src/valueobject/Attribute.php
@@ -52,7 +52,7 @@ class Attribute
         return $this->description;
     }
 
-    public function getPriceForValue(string $value): int
+    public function getPriceForValue(string $value): float
     {
         return $this->valueprice[$value];
     }
